requirements.txt
keyboard==0.13.5
numpy==2.0.1
opencv_python==4.10.0.84
Pillow==10.4.0
PyAutoGUI==0.9.54
pystray==0.19.5
